Mosin trees!
The trees that give you the Mosin Nagant sniper!

Works properly!

(Also, I am coming back to this project full time, because surviv.io officially is going to be deleted, and we need to create our OpenSource version faster, for people of the public to play!)
LET THE .IO GAMES RULE THE GAMING INDUSTRY!